Introduce BootACMR
