+++ spring-cloud-dataflow-configuration-metadata/src/main/java/org/springframework/cloud/dataflow/configuration/metadata/ApplicationConfigurationMetadataResolverAutoConfiguration.java
@@ -31,6 +31,6 @@ public class ApplicationConfigurationMetadataResolverAutoConfiguration {
 	@Bean
 	@ConditionalOnMissingBean(ApplicationConfigurationMetadataResolver.class)
 	public ApplicationConfigurationMetadataResolver metadataResolver() {
-		return new ApplicationConfigurationMetadataResolver();
+		return new BootApplicationConfigurationMetadataResolver();
 	}
 }
